Fix resizing issue; #78
The is_only_window function should in fact not check whether or not it
is the only window but rather, if there is no other vim window in the
direction towards which we want to resize.

For instance, when resizing in a horizontal direction with multiple
horizontal splits open, we should resize the tmux pane instead of -
fruitlessly - trying to resize the vim window.